[desktop] add spell checking, close #1069
adds spell check & correction capabilities to the context menu,
including the ability to change the checked language.

dictionaries are pulled from <updateUrl>/../Dictionaries/ except
for custom builds. in that case, the downloader is pointed to the
cache directory for dictionaries:

~/.config/tutanota-desktop/Dictionaries/ on linux
%APPDATA%\tutanota-desktop\Dictionaries on windows
~/Library/Application Support/tutanota-desktop/Dictionaries/ on mac

this results in no dictionaries being downloaded and the requirement to
place them into that cache dir manually.

the dist.js build script gained a --get-dicts flag, which will
result in the dictionaries released with the used electron version being
downloaded from

https://github.com/electron/electron/releases/
download/v${ersion}/hunspell_dictionaries.zip

and extracted into a directory next to the given --out-dir
or build/dist/Dictionaries during the build.